removed undeployed version of allowTokens contracts
