--- federator/src/contracts/IAllowTokensV2.ts
@@ -1,79 +1,0 @@
-import CustomError from '../lib/CustomError';
-import { VERSIONS } from './Constants';
-
-interface GetLimitsParams {
-  tokenAddress: string;
-}
-
-export class IAllowTokensV2 {
-  allowTokensContract: any;
-  mapTokenInfoAndLimits: any;
-  chainId: number;
-
-  constructor(allowTokensContract, chainId: number) {
-    this.allowTokensContract = allowTokensContract;
-    this.mapTokenInfoAndLimits = {};
-    this.chainId = chainId;
-  }
-
-  getVersion() {
-    return VERSIONS.V2;
-  }
-
-  async getConfirmations() {
-    const promises = [];
-    promises.push(this.getSmallAmountConfirmations());
-    promises.push(this.getMediumAmountConfirmations());
-    promises.push(this.getLargeAmountConfirmations());
-    const result = await Promise.all(promises);
-    return {
-      smallAmountConfirmations: result[0],
-      mediumAmountConfirmations: result[1],
-      largeAmountConfirmations: result[2],
-    };
-  }
-
-  async getSmallAmountConfirmations() {
-    try {
-      return this.allowTokensContract.methods.smallAmountConfirmations().call();
-    } catch (err) {
-      throw new CustomError(`Exception getSmallAmountConfirmations at AllowTokens Contract`, err);
-    }
-  }
-
-  async getMediumAmountConfirmations() {
-    try {
-      return this.allowTokensContract.methods.mediumAmountConfirmations().call();
-    } catch (err) {
-      throw new CustomError(`Exception getMediumAmountConfirmations at AllowTokens Contract`, err);
-    }
-  }
-
-  async getLargeAmountConfirmations() {
-    try {
-      return this.allowTokensContract.methods.largeAmountConfirmations().call();
-    } catch (err) {
-      throw new CustomError(`Exception getLargeAmountConfirmations at AllowTokens Contract`, err);
-    }
-  }
-
-  async getLimits(objParams: GetLimitsParams) {
-    try {
-      let result = this.mapTokenInfoAndLimits[objParams.tokenAddress];
-      if (!result) {
-        const infoAndLimits = await this.allowTokensContract.methods.getInfoAndLimits(objParams.tokenAddress).call();
-        result = {
-          allowed: infoAndLimits.info.allowed,
-          mediumAmount: infoAndLimits.limit.mediumAmount,
-          largeAmount: infoAndLimits.limit.largeAmount,
-        };
-        if (result.allowed) {
-          this.mapTokenInfoAndLimits[objParams.tokenAddress] = result;
-        }
-      }
-      return result;
-    } catch (err) {
-      throw new CustomError(`Exception getInfoAndLimits at AllowTokens Contract`, err);
-    }
-  }
-}
